RNS: use double for chemstatus

+++ RNS/src/Src_3d/RNS_F.f90
@@ -194,7 +194,7 @@ subroutine rns_advchem(lo,hi,U,U_l1,U_l2,U_l3,U_h1,U_h2,U_h3,  &
   integer, intent(in) ::  U_l1, U_l2, U_l3, U_h1, U_h2, U_h3, &
        st_l1,st_l2,st_l3,st_h1,st_h2,st_h3
   double precision, intent(inout) :: U(U_l1:U_h1,U_l2:U_h2,U_l3:U_h3,NVAR)
-  integer, intent(inout) :: st(st_l1:st_h1,st_l2:st_h2,st_l3:st_h3)
+  double precision, intent(inout) :: st(st_l1:st_h1,st_l2:st_h2,st_l3:st_h3)
   double precision, intent(in) :: dt
 
   integer :: Ulo(3), Uhi(3), stlo(3), sthi(3)
@@ -219,7 +219,7 @@ subroutine rns_advchem2(lo,hi,U,U_l1,U_l2,U_l3,U_h1,U_h2,U_h3, &
        st_l1,st_l2,st_l3,st_h1,st_h2,st_h3
   integer, intent(in) ::  Up_l1, Up_l2, Up_l3, Up_h1, Up_h2, Up_h3
   double precision, intent(inout) :: U(U_l1:U_h1,U_l2:U_h2,U_l3:U_h3,NVAR)
-  integer, intent(inout) :: st(st_l1:st_h1,st_l2:st_h2,st_l3:st_h3)
+  double precision, intent(inout) :: st(st_l1:st_h1,st_l2:st_h2,st_l3:st_h3)
   double precision, intent(in   ) :: Up(Up_l1:Up_h1,Up_l2:Up_h2,Up_l3:Up_h3,NVAR)
   double precision, intent(in) :: dt
 
@@ -250,7 +250,7 @@ subroutine rns_dUdt_chem(lo,hi, &
        U_l3, U_h3, Ut_l3, Ut_h3, st_l1,st_l2,st_l3,st_h1,st_h2,st_h3
   double precision, intent(in ) ::  U( U_l1: U_h1, U_l2: U_h2, U_l3: U_h3,NVAR)
   double precision, intent(out) :: Ut(Ut_l1:Ut_h1,Ut_l2:Ut_h2,Ut_l3:Ut_h3,NVAR)
-  integer, intent(inout) :: st(st_l1:st_h1,st_l2:st_h2,st_l3:st_h3)
+  double precision, intent(inout) :: st(st_l1:st_h1,st_l2:st_h2,st_l3:st_h3)
 
   integer :: Ulo(3), Uhi(3), Utlo(3), Uthi(3), stlo(3), sthi(3)
